pyproject.toml
[project]
name = "fin-restructure-assistant"
version = "0.1.0"
description = "Financial Restructuring Assistant using FastAPI and GenAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-storage-blob>=12.27.0",
    "coverage>=7.11.0",
    "fastapi[standard]>=0.119.1",
    "httpx>=0.28.1",
    "openai>=2.6.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "fpdf2>=2.7.10",
    "ruff>=0.14.1",
    "uvicorn>=0.30.0",
    "fpdf>=1.7.2",
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term --cov-report=html"

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
